requirements.txt
alabaster==0.7.12
asgiref==3.3.4
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==5.0.5
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.11.2
django-datetime-widget==0.9.3
django-extensions==3.1.2
django-heroku==0.3.1
django-on-heroku==1.1.2
django-shell-plus==1.1.7
docutils==0.17
gitdb==4.0.7
GitPython==3.1.14
gunicorn==20.1.0
heroku==0.1.4
idna==3.1
imagesize==1.2.0
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
parso==0.8.2
pguard==0.4.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
psycopg2-binary==2.8.6
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
whitenoise==5.2.0
